% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{assert_awr_jars}
\alias{assert_awr_jars}
\title{Asserts if AWS SDK for Java jar files are available}
\usage{
assert_awr_jars()
}
\value{
invisible \code{TRUE} on success, otherwise error
}
\description{
Asserts if AWS SDK for Java jar files are available
}
