% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{AWR-package}
\alias{AWR-package}
\title{Installing AWS SDK jars for R}
\description{
Downloads and installs the AWS SDK for Java jars from the \code{ARW.jars} packages, which also adds to the Java classpath to be easily used in other R packages.
}
\references{
\url{https://aws.amazon.com/sdk-for-java}
}

