% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_awr_jars}
\alias{check_awr_jars}
\title{Checks if AWS SDK for Java jar files are available}
\usage{
check_awr_jars()
}
\value{
\code{TRUE} on success, informative message as a string on error
}
\description{
Checks if AWS SDK for Java jar files are available
}
