% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{AWR-package}
\alias{AWR-package}
\title{Making the AWS Java SDK JAR classes available in R}
\description{
This R package makes the \code{jar} files of the AWS SDK for Java
available to be used in downstream R packages. Please note the
installation instructions for the System Requirements in the
\code{README.md}.
}
\examples{
\dontrun{
library(rJava)
client <- .jnew("com.amazonaws.services.s3.AmazonS3Client")
kc$getS3AccountOwner()$getDisplayName()
}
}
\references{
\url{https://aws.amazon.com/sdk-for-java}
}
